Location Stuff
Got location - added example businessess with some logic that needs to be changed for attaching and displaying with certain cards.

TODO

Sign up for affiliate API's + Better implement local business suggestions / cards.

Is this supposed to be something that you stay at home and play and answer questions or something to come up with an idea to get out and do? Do we need to support direct messaging?